Rules update and new Terms of Service/Privacy Policy
This commit includes two big changes:
Better rule phrasing. This change groups the rules in "rules" and "best practices", and changes some sentences to be more precise and inclusive. The best practices make a more explicit case for decent and accessible behaviour, and mirror how our instance is run in practice already.

The second change concerns the legalese and brings us new Terms of Service. The terms provided by Mastodon consist of only a Privacy Policy, which is not accurate for how we run chaos.social, so we wrote our own.

Thanks go to the administrators of bsd.network, who have allowed us to copy/steal/be inspired by parts of their terms of service.